requirements.txt
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==1.0.0
SQLAlchemy==1.3.3
Werkzeug==1.0.1

 
 

 